Rearrange link module
