+++ src/link/biblio/mod.rs
@@ -1,6 +1,8 @@
+pub mod manager;
+mod source;
+
 use kuchiki::NodeRef;
 
-use super::source::BiblioEntrySource;
 use crate::html::{self, Attr};
 
 #[derive(Debug, Default, Clone)]
@@ -32,22 +34,5 @@ impl BiblioEntry {
         ));
 
         dd_el
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct BiblioEntryManager {
-    pub biblio_entry_source: BiblioEntrySource,
-}
-
-impl BiblioEntryManager {
-    pub fn new() -> Self {
-        BiblioEntryManager {
-            biblio_entry_source: BiblioEntrySource::new("spec-data"),
-        }
-    }
-
-    pub fn get_biblio_entry(&mut self, spec: &str) -> BiblioEntry {
-        self.biblio_entry_source.fetch_biblio_entry(spec)
     }
 }
