+++ src/spec.rs
@@ -10,8 +10,9 @@ use crate::fix::{self, CodeSpanManager};
 use crate::heading;
 use crate::html;
 use crate::line::Line;
-use crate::link::biblio::BiblioEntryManager;
-use crate::link::reference::{Reference, ReferenceManager};
+use crate::link::biblio::manager::BiblioEntryManager;
+use crate::link::reference::manager::ReferenceManager;
+use crate::link::reference::Reference;
 use crate::link::{self, biblio::BiblioEntry, dfn};
 use crate::markdown;
 use crate::metadata::{self, Metadata};
